Move provider tests into provider package
